requirements.txt
requests
pytz
scipy
numpy
mysql-connector
